Finished C# range implementation

+++ integration-tests/TestAPI-bindings/c-sharp/Cricut.TestAPI.Tests/PrimitiveTests.cs
@@ -20,8 +20,8 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(ulong.MinValue, Primitives.MinUInt64);
             Assert.Equal(ulong.MaxValue, Primitives.MaxUInt64);
             Assert.Equal(0u, Primitives.ZeroUInt);
-            Assert.Equal(ulong.MinValue, Primitives.MinUInt);
-            Assert.Equal(ulong.MaxValue, Primitives.MaxUInt);
+            Assert.Equal(nuint.MinValue, Primitives.MinUInt);
+            Assert.Equal(nuint.MaxValue, Primitives.MaxUInt);
             Assert.Equal(0, Primitives.ZeroInt8);
             Assert.Equal(sbyte.MinValue, Primitives.MinInt8);
             Assert.Equal(sbyte.MaxValue, Primitives.MaxInt8);
@@ -35,8 +35,8 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(long.MinValue, Primitives.MinInt64);
             Assert.Equal(long.MaxValue, Primitives.MaxInt64);
             Assert.Equal(0, Primitives.ZeroInt);
-            Assert.Equal(long.MinValue, Primitives.MinInt);
-            Assert.Equal(long.MaxValue, Primitives.MaxInt);
+            Assert.Equal(nint.MinValue, Primitives.MinInt);
+            Assert.Equal(nint.MaxValue, Primitives.MaxInt);
             Assert.Equal(0.0f, Primitives.ZeroFloat);
             Assert.Equal(-float.MaxValue, Primitives.MinFloat);
             Assert.Equal(float.MaxValue, Primitives.MaxFloat);
@@ -52,12 +52,12 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(new ushort[] { (ushort)0, ushort.MinValue, ushort.MaxValue }, Primitives.ManyUInt16);
             Assert.Equal(new uint[] { 0u, uint.MinValue, uint.MaxValue }, Primitives.ManyUInt32);
             Assert.Equal(new ulong[] { 0u, ulong.MinValue, ulong.MaxValue }, Primitives.ManyUInt64);
-            Assert.Equal(new ulong[] { 0u, ulong.MinValue, ulong.MaxValue }, Primitives.ManyUInt);
+            Assert.Equal(new nuint[] { 0u, nuint.MinValue, nuint.MaxValue }, Primitives.ManyUInt);
             Assert.Equal(new sbyte[] { 0, sbyte.MinValue, sbyte.MaxValue }, Primitives.ManyInt8);
             Assert.Equal(new short[] { 0, short.MinValue, short.MaxValue }, Primitives.ManyInt16);
             Assert.Equal(new int[] { 0, int.MinValue, int.MaxValue }, Primitives.ManyInt32);
             Assert.Equal(new long[] { 0, long.MinValue, long.MaxValue }, Primitives.ManyInt64);
-            Assert.Equal(new long[] { 0, long.MinValue, long.MaxValue }, Primitives.ManyInt);
+            Assert.Equal(new nint[] { 0, nint.MinValue, nint.MaxValue }, Primitives.ManyInt);
             Assert.Equal(new float[] { 0.0f, -float.MaxValue, float.MaxValue }, Primitives.ManyFloat);
             Assert.Equal(new double[] { 0.0, -double.MaxValue, double.MaxValue }, Primitives.ManyDouble);
         }
@@ -69,12 +69,12 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(new ushort?[] { null, 0, ushort.MinValue, ushort.MaxValue }, Primitives.ManyMaybeUInt16);
             Assert.Equal(new uint?[] { null, 0, uint.MinValue, uint.MaxValue }, Primitives.ManyMaybeUInt32);
             Assert.Equal(new ulong?[] { null, 0, ulong.MinValue, ulong.MaxValue }, Primitives.ManyMaybeUInt64);
-            Assert.Equal(new ulong?[] { null, 0, ulong.MinValue, ulong.MaxValue }, Primitives.ManyMaybeUInt);
+            Assert.Equal(new nuint?[] { null, 0, nuint.MinValue, nuint.MaxValue }, Primitives.ManyMaybeUInt);
             Assert.Equal(new sbyte?[] { null, 0, sbyte.MinValue, sbyte.MaxValue }, Primitives.ManyMaybeInt8);
             Assert.Equal(new short?[] { null, 0, short.MinValue, short.MaxValue }, Primitives.ManyMaybeInt16);
             Assert.Equal(new int?[] { null, 0, int.MinValue, int.MaxValue }, Primitives.ManyMaybeInt32);
             Assert.Equal(new long?[] { null, 0, long.MinValue, long.MaxValue }, Primitives.ManyMaybeInt64);
-            Assert.Equal(new long?[] { null, 0, long.MinValue, long.MaxValue }, Primitives.ManyMaybeInt);
+            Assert.Equal(new nint?[] { null, 0, nint.MinValue, nint.MaxValue }, Primitives.ManyMaybeInt);
             Assert.Equal(new float?[] { null, 0.0f, -float.MaxValue, float.MaxValue }, Primitives.ManyMaybeFloat);
             Assert.Equal(new double?[] { null, 0.0, -double.MaxValue, double.MaxValue }, Primitives.ManyMaybeDouble);
         }
@@ -96,8 +96,8 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(ulong.MinValue, Primitives.EchoUInt64(ulong.MinValue));
             Assert.Equal(ulong.MaxValue, Primitives.EchoUInt64(ulong.MaxValue));
             Assert.Equal(0u, Primitives.EchoUInt(0u));
-            Assert.Equal(ulong.MinValue, Primitives.EchoUInt(ulong.MinValue));
-            Assert.Equal(ulong.MaxValue, Primitives.EchoUInt(ulong.MaxValue));
+            Assert.Equal(nuint.MinValue, Primitives.EchoUInt(nuint.MinValue));
+            Assert.Equal(nuint.MaxValue, Primitives.EchoUInt(nuint.MaxValue));
             Assert.Equal(0, Primitives.EchoInt8(0));
             Assert.Equal(sbyte.MinValue, Primitives.EchoInt8(sbyte.MinValue));
             Assert.Equal(sbyte.MaxValue, Primitives.EchoInt8(sbyte.MaxValue));
@@ -111,8 +111,8 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(long.MinValue, Primitives.EchoInt64(long.MinValue));
             Assert.Equal(long.MaxValue, Primitives.EchoInt64(long.MaxValue));
             Assert.Equal(0, Primitives.EchoInt(0));
-            Assert.Equal(long.MinValue, Primitives.EchoInt(long.MinValue));
-            Assert.Equal(long.MaxValue, Primitives.EchoInt(long.MaxValue));
+            Assert.Equal(nint.MinValue, Primitives.EchoInt(nint.MinValue));
+            Assert.Equal(nint.MaxValue, Primitives.EchoInt(nint.MaxValue));
             Assert.Equal(0.0f, Primitives.EchoFloat(0.0f));
             Assert.Equal(-float.MaxValue, Primitives.EchoFloat(-float.MaxValue));
             Assert.Equal(float.MaxValue, Primitives.EchoFloat(float.MaxValue));
@@ -144,8 +144,8 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(ulong.MaxValue, Primitives.MaybeEchoUInt64(ulong.MaxValue));
             Assert.Null(Primitives.MaybeEchoUInt(null));
             Assert.Equal(0u, Primitives.MaybeEchoUInt(0u));
-            Assert.Equal(ulong.MinValue, Primitives.MaybeEchoUInt(ulong.MinValue));
-            Assert.Equal(ulong.MaxValue, Primitives.MaybeEchoUInt(ulong.MaxValue));
+            Assert.Equal(nuint.MinValue, Primitives.MaybeEchoUInt(nuint.MinValue));
+            Assert.Equal(nuint.MaxValue, Primitives.MaybeEchoUInt(nuint.MaxValue));
             Assert.Null(Primitives.MaybeEchoInt8(null));
             Assert.Equal((sbyte)0, Primitives.MaybeEchoInt8((sbyte)0));
             Assert.Equal(sbyte.MinValue, Primitives.MaybeEchoInt8(sbyte.MinValue));
@@ -164,8 +164,8 @@ namespace Cricut.TestAPI.Tests {
             Assert.Equal(long.MaxValue, Primitives.MaybeEchoInt64(long.MaxValue));
             Assert.Null(Primitives.MaybeEchoInt(null));
             Assert.Equal(0, Primitives.MaybeEchoInt(0));
-            Assert.Equal(long.MinValue, Primitives.MaybeEchoInt(long.MinValue));
-            Assert.Equal(long.MaxValue, Primitives.MaybeEchoInt(long.MaxValue));
+            Assert.Equal(nint.MinValue, Primitives.MaybeEchoInt(nint.MinValue));
+            Assert.Equal(nint.MaxValue, Primitives.MaybeEchoInt(nint.MaxValue));
             Assert.Null(Primitives.MaybeEchoFloat(null));
             Assert.Equal(0.0f, Primitives.MaybeEchoFloat(0.0f));
             Assert.Equal(-float.MaxValue, Primitives.MaybeEchoFloat(-float.MaxValue));
